Moved the function from scrape directory to scraper.py
